events: Don't ignore first mouse motion after warp
Since commit cf092eca "mouse: Save initial position yet even if xrel
and yrel are 0", the first mouse event after warping the mouse would
always be processed, even if it didn't change any state. However, this
could cause an event queue overflow in the Daemon game engine used in
Unvanquished, which warps the mouse in response to each SDL_MOUSEMOTION
event, causing the next mouse motion to generate a SDL_MOUSEMOTION event
even if its displacement is zero, which warps the mouse again, and so on.

Commit 33185907 "Fix relative mouse input for Unvanquished" attempted
to fix this, but may have done so with a bigger hammer than was really
intended: it skipped the entire rest of the function for the first
mouse motion event after a pointer warp, not just the part that
delivered the SDL_MOUSEMOTION to engine code. This caused odd results
in Team Fortress 2 and Portal 2, in which fast mouse movement would
periodically jump back to a previous position.

Instead of skipping the whole body of the function, just skip the part
where the SDL_MOUSEMOTION event is delivered to application code.
This lets SDL apply the rest of its internal logic, in particular
SDL_WINDOW_MOUSE_CAPTURE and updating the last_x and last_y coordinates.

This partially reverts commit 331859079674465a39b24f32a6a113959601dca3.

Bug: https://github.com/libsdl-org/SDL/issues/5569
Signed-off-by: Simon McVittie <smcv@collabora.com>