README-migration: Suggest a cross-compilation-friendly Makefile snippet
Hard-coding the pkg-config executable name is a very, very common
portability issue, and many games and other programs can be made
cross-compilation-friendly without any other changes by accepting
PKG_CONFIG (and other compilation tools like CC or CXX) from the
environment. The quakespasm Quake engine is a good example of this
technique.

The Linux-kernel-style convention of invoking ${CROSS_COMPILE}gcc,
${CROSS_COMPILE}pkg-config and so on, with CROSS_COMPILE left empty
for native builds and set to something like i686-w64-mingw32- (with
the trailing dash!) for cross-builds, is a simple way to make a Makefile
build system cross-friendly, so nudge developers towards that.

Signed-off-by: Simon McVittie <smcv@collabora.com>